Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
    "zaproxy>=0.4.0",
]
